added

+++ src/app/(sikundi)/sikundi-admin/(collections)/graphic/_actions/update/index.ts
@@ -5,6 +5,7 @@ import GraphicSchema, { GraphicSchemaType } from './schema'
 import ErrorHandler from '@sikundi/lib/server/utils/ErrorHandler'
 import { prisma } from '@sikundi/lib/server/utils/prisma'
 import { revalidatePath } from 'next/cache'
+import getPermission from '@sikundi/lib/server/utils/getPermission'
 
 const statusFromActions = {
     draft: "drafted",
@@ -16,6 +17,26 @@ const statusFromActions = {
 export default async function POST(data: GraphicSchemaType) {
     return (await ErrorHandler(data, GraphicSchema, async ({createdBy, graphicsUrl, language, push, action, id, ...data}:GraphicSchemaType) => {
         const user = await getUser()
+
+        const permission = await getPermission({
+            graphic: [
+                "draft",
+                "delete",
+                "soft_delete",
+                "publish",
+                "pending"
+            ]
+        })
+
+        if(!permission?.graphic?.[String(action)]) {
+            throw({
+                notification: {
+                    title: 'Authorization Error',
+                    description: `You are not allowed to ${action} graphics.`,
+                    variant: "destructive"
+                }
+            })
+        }
 
         if (action === "delete") {
             const graphic = await prisma.graphic.delete({
@@ -36,7 +57,7 @@ export default async function POST(data: GraphicSchemaType) {
             })
         }
 
-
+        
         const graphic = await prisma.graphic.update({
             data: {
                 ...data,
