+++ src/app/(sikundi)/sikundi-admin/(collections)/photo/_actions/update/index.ts
@@ -5,6 +5,7 @@ import PhotoSchema, { PhotoSchemaType } from './schema'
 import ErrorHandler from '@sikundi/lib/server/utils/ErrorHandler'
 import { prisma } from '@sikundi/lib/server/utils/prisma'
 import { revalidatePath } from 'next/cache'
+import getPermission from '@sikundi/lib/server/utils/getPermission'
 
 const statusFromActions = {
     draft: "drafted",
@@ -16,6 +17,25 @@ const statusFromActions = {
 export default async function POST(data: PhotoSchemaType) {
     return (await ErrorHandler(data, PhotoSchema, async ({createdBy, featureImageUrl, language, push, action, id, ...data}:PhotoSchemaType) => {
         const user = await getUser()
+        const permission = await getPermission({
+            photo: [
+                "draft",
+                "delete",
+                "soft_delete",
+                "publish",
+                "pending"
+            ]
+        })
+
+        if(!permission?.photo?.[String(action)]) {
+            throw({
+                notification: {
+                    title: 'Authorization Error',
+                    description: `You are not allowed to ${action} photos.`,
+                    variant: "destructive"
+                }
+            })
+        }
 
         if (action === "delete") {
             const photo = await prisma.photo.delete({
